% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casecohort_secondary.R
\docType{data}
\name{casecohort_data_secondary}
\alias{casecohort_data_secondary}
\title{Data example for the secondary analysis in case-cohort design}
\format{A data frame with 1000 rows and 15 columns: \describe{
  \item{subj_ind}{An indicator variable for each subject: 1 = SRS, 2 =
  supplemental cases, 0 = NVsample} \item{T}{observation time for failure
  outcome} \item{Delta}{event indicator} \item{Y2}{a continuous secondary
  outcome}\item{X}{expensive exposure} \item{Z11}{first covariate in the
  linear regression model} \item{Z12}{second covariate in the linear
  regression model}\item{Z13}{third covariate in the linear regression model}
  \item{Z14}{fourth covariate in the linear regression model}\item{Z21}{first
  covariate in the Cox model} \item{Z22}{second covariate in the Cox
  model}\item{Z23}{third covariate in the Cox model} \item{Z31}{first
  covariate that is related to the conditional distribution of X given other
  covariates}\item{Z32}{second covariate that is related to the conditional
  distribution} \item{Z33}{thid covariate that is related to the conditional
  distribution} }}
\source{
A simulated data set
}
\usage{
casecohort_data_secondary
}
\description{
Data example for the secondary analysis in case-cohort design
}
\keyword{datasets}
