% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_el.R
\name{L.el}
\alias{L.el}
\title{Calculates the profiled empirical likelihood}
\usage{
L.el(ys, mean, sd, pi.fn, N, log = FALSE, R = 100, tol = 1)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{mean}{vector of means of dependent variable}

\item{sd}{vector of standard deviations of dependent variable}

\item{pi.fn}{a function taking a single argument y and returning the probability
of selection}

\item{N}{the population size}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}

\item{R}{number of points used in quadrature calculation of integral}

\item{tol}{a factor controlling error-tolerance in integration (defaults to 1)}
}
\value{
a quantity proportional to the density of ys (or the log thereof)
}
\description{
This function calculates the profiled empirical likelihood (or rather something proportional to it)
}
\details{
The dependent variable is assumed to be normally distributed
}
\examples{
data(population_example)
L.el(ys=sample.example$y,mean=5+0.5*sample.example$x,sd=1,log=TRUE,N=1000,
pi.fn=function(y,log=FALSE){
out <- pmin(1,0.05*sqrt(0.1+0.9*(y-5)^2/1))
if(log) return(log(out))
if(!log) return(out)})
}
