% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_est.R
\name{L.est}
\alias{L.est}
\title{Calculates the estimated likelihood for a regression model under a general sample design}
\usage{
L.est(beta, sd, ys, xs, N, p.s, specs = NULL, log = FALSE, pi.s, R = 100,
  all.resamp.x, all.errors.y, verbose = FALSE)
}
\arguments{
\item{beta}{vector of regression coefficients}

\item{sd}{error standard deviation}

\item{ys}{vector of sample values of the dependent variable}

\item{xs}{matrix of sample covariate values of dimension n by p
where n is the sample size (length(ys))}

\item{N}{the population size}

\item{p.s}{A function defining the ODS sample design.
It must take arguments ys, yr, log and stuff, and return
the probability (or log of the probability) that a particular
sample was selected.}

\item{specs}{An object containing detailed specifications of the design.}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}

\item{pi.s}{the probabilities of selection of the sample units}

\item{R}{the number of simulations used in the monte carlo approximation
of the likelihood}

\item{all.resamp.x}{Optional. An R by N-n integer-valued matrix
whose rows are SRSWR resamples from 1:n}

\item{all.errors.y}{Optional. An R by N-n real-valued matrix
of standard normal realisations.}

\item{verbose}{If TRUE, the function outputs information to the console.}
}
\value{
The likelihood or log-likelihood.
}
\description{
This function calculates the estimated likelihood
}
\details{
Add some details later.
}
\examples{
data(population_example)
L.est(beta=c(4,1),N=1000,sd=0.8,ys=sample.example$y,xs=cbind(1,sample.example$y),
p.s=p.s.ushaped,specs=c(n0=10,tuner=0.1,return.what=1),log=TRUE,
pi.s=sample.example$pi)
}
