% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_modZhou.R
\name{L.modZhou}
\alias{L.modZhou}
\title{Calculates a modified Zhou conditional likelihood}
\usage{
L.modZhou(ys, mean, sd, cutoffs, log = FALSE)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{mean}{vector of means of dependent variable}

\item{sd}{vector of standard deviations of dependent variable}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}
}
\value{
likelihood or log-likelihood conditional on stratum memberships
}
\description{
This function calculates a conditional likelihood
which is a modification of the one in Zhou et al. (2002).
}
\details{
The dependent variable is assumed to be normally distributed
}
\examples{
data(population_example)
L.modZhou(ys=sample.example$y,mean=5+0.5*sample.example$x,sd=1,log=TRUE,cutoffs=c(1,2,3))
}

