% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_WZ.R
\name{L.WZ}
\alias{L.WZ}
\title{Calculates Weaver and Zhou likelihood}
\usage{
L.WZ(ys, yr, xs, beta, sd, cutoffs, log = F)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{yr}{vector of non-sample values of the dependent variable}

\item{xs}{matrix of sample covariate values of dimension n by p
where n is the sample size (length(ys))}

\item{beta}{vector of regression coefficients}

\item{sd}{error standard deviation}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}
}
\value{
likelihood or log-likelihood conditional on stratum memberships
}
\description{
This function calculates the likelihood from Weaver and Zhou (2012)
}
\details{
The dependent variable is assumed to be normally distributed
}
\examples{
data(population_example)
L.WZ(ys=sample.example$y,yr=population.example$y[population.example$s.indicators==0],
xs=cbind(1,sample.example$x),beta=c(1,1),sd=1,cutoffs=c(1,2,3),log=TRUE)
}

