% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Zhou_2x.R
\name{L.Zhou.2x}
\alias{L.Zhou.2x}
\title{Calculates the estimated likelihood for a regression model under a general sample design
assuming one continuous and one binary covariate}
\usage{
L.Zhou.2x(beta, sd, prob.x2, mu0.x1, mu1.x1, sd.x1, ys, x1s, x2s, cutoffs,
  in.s0, log = FALSE)
}
\arguments{
\item{beta}{vector of regression coefficients}

\item{sd}{error standard deviation}

\item{prob.x2}{the parameter of x2's assumed bernoulli distribution}

\item{mu0.x1}{the expected value of x1 conditional on x2=0}

\item{mu1.x1}{the expected value of x1 conditional on x2=1}

\item{sd.x1}{the SD of x1 conditional on x2}

\item{ys}{vector of sample values of the dependent variable}

\item{x1s}{sample values for covariate 1 (continuous)}

\item{x2s}{sample values for covariate 2 (binary)}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{in.s0}{A vector equal to 1 for units selected in the SRSWOR and to 0 for those
in the stratified SRSWOR. Default is that there are no SRSWOR units.}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}
}
\value{
The likelihood or log-likelihood.
}
\description{
This function calculates the likelihood.
}
\details{
Add some details later.
}
\examples{
data(population_example)
eg.ybar.pi <- sum(sample.example$y/sample.example$pi) / sum(1/sample.example$pi)
eg.s2.pi <- sum((sample.example$y-eg.ybar.pi)^2/sample.example$pi)/sum(1/sample.example$pi)
L.Zhou.2x(beta=c(1,0.5,-0.5),sd=1,prob.x2=0.45,mu0.x1=0,mu1.x1=0,sd.x1=1,
       ys=sample.example$y,x1s=sample.example$x1,x2s=sample.example$x2,
       cutoffs=c(0,2),log=TRUE)
}
