% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_MLE_2x_poisson_logistic.R
\name{lm.MLE.2x.poisson.logistic}
\alias{lm.MLE.2x.poisson.logistic}
\title{Calculates maximum likelihood estimators for linear regression of complex survey data}
\usage{
lm.MLE.2x.poisson.logistic(ys, x1s, x2s, pi.s, pi1, pi0, start.beta)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{x1s}{sample values for covariate 1 (continuous)}

\item{x2s}{sample values for covariate 2 (binary)}

\item{pi.s}{n-vector (where n=length(y)) of probabilities of selection for the sample units}

\item{pi1}{The probability of selection when Y=1}

\item{pi0}{The probability of selection when Y=0}

\item{start.beta}{vector containing (b0,b1,b2) start values}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. No variance estimates provided.
}
\description{
This function fits a linear model to complex sample data by maximum likelihood
}
\details{
Add some details later.
}
\examples{
data(population_example)
eg.ybar.pi <- sum(sample.example$y/sample.example$pi) / sum(1/sample.example$pi)
eg.s2.pi <- sum((sample.example$y-eg.ybar.pi)^2/sample.example$pi)/sum(1/sample.example$pi)
eg.pi.fn <- function(y){ out <- sqrt( 0.1 + 0.9*(y-eg.ybar.pi)^2/eg.s2.pi )
out <- out * mean(sample.example$pi) /
       mean(sqrt( 0.1 + 0.9*(sample.example$y-eg.ybar.pi)^2/eg.s2.pi ))
out  }
lm.MLE.2x.poisson(ys=sample.example$y,x1s=sample.example$x1,x2s=sample.example$x2,pi.fn=eg.pi.fn,
       pi.s=sample.example$pi,R=10)
}
