% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_MLE_2x_poisson_strat.R
\name{lm.MLE.2x.poisson.strat}
\alias{lm.MLE.2x.poisson.strat}
\title{Calculates maximum likelihood estimators for linear regression of complex survey data}
\usage{
lm.MLE.2x.poisson.strat(ys, x1s, x2s, pi.s, pi.h, cutoffs, start.beta, start.sd)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{x1s}{sample values for covariate 1 (continuous)}

\item{x2s}{sample values for covariate 2 (binary)}

\item{pi.s}{n-vector (where n=length(y)) of probabilities of selection for the sample units}

\item{pi.h}{H-vector (where H=number of strata) of probs of selection in each stratum}

\item{cutoffs}{(H-1)-vector of cutoffs defining strata}

\item{start.beta}{vector containing (b0,b1,b2) start values}

\item{start.sd}{scalar start value for sd of Y given X}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. No variance estimates provided.
}
\description{
This function fits a linear model to complex sample data by maximum likelihood
}
\details{
Add some details later.
}
\examples{
data(population_example)
lm.MLE.2x.poisson.strat(ys=sample.example$y,x1s=sample.example$x1,x2s=sample.example$x2,
pi.h=c(0.1,0.05,0.1),cutoffs=c(0,2),
pi.s=0.1-0.05*(sample.example$y>=0)*(sample.example$y<=2))
}
