% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_MLE.R
\name{lm.MLE}
\alias{lm.MLE}
\title{Calculates maximum likelihood estimators for linear regression of complex survey data}
\usage{
lm.MLE(ys, xs, p.s, specs, pi.s, N, R)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{xs}{matrix of covariate values (number of rows must equal n where n is length of ys).
Intercept is not assumed, so xs should contain a column of 1s if you want an intercept in the model.}

\item{p.s}{A function defining the ODS sample design.
It must take arguments ys, yr, log and stuff, and return
the probability (or log of the probability) that a particular
sample was selected.}

\item{specs}{An object containing detailed specifications of the design.}

\item{pi.s}{n-vector (where n=length(y)) of probabilities of selection for the sample units}

\item{N}{the population size}

\item{R}{the number of simulations used in the monte carlo approximation
of the likelihood}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. No variance estimates provided.
}
\description{
This function fits a linear model to complex sample data by maximum likelihood
}
\details{
Add some details later.
}
\examples{
data(population_example)
lm.MLE(ys=sample.example$y,xs=cbind(1,sample.example$x),p.s=p.s.ushaped,
       specs=c(n0=10,tuner=0.1,return.what=1),pi.s=sample.example$pi,N=100,R=10)
}

