% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_samplik_2x_poisson_strat.R
\name{lm.samplik.2x.poisson.strat}
\alias{lm.samplik.2x.poisson.strat}
\title{Calculates maximum sample likelihood estimators for linear regression}
\usage{
lm.samplik.2x.poisson.strat(ys, xs, pi.fn, pi.h, cutoffs,
  estimate.var = FALSE)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{xs}{matrix of covariate values (number of rows must equal n where n is length of ys).
Intercept is not assumed, so xs should contain a column of 1s if you want an intercept in the model.}

\item{pi.fn}{a function taking a single argument y and returning the probability
of selection}

\item{pi.h}{selection probability in each stratum (H-vector)}

\item{cutoffs}{vector of H-1 cutoffs on Y defining the strata}

\item{estimate.var}{If TRUE, variance is estimated using the observed information. Defaults to FALSE}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. If estimate.var=TRUE then variance estimates are provided.
}
\description{
This function fits a linear model to complex sample data by maximum sample likelihood
}
\details{
See Krieger and Pfefferman's paper.
The dependent variable is assumed to be normally distributed.
}
\examples{
data(population_example)
lm.samplik(ys=sample.example$y,xs=cbind(1,sample.example$x),
pi.fn=function(y,log=FALSE){
out <- sqrt(0.1+0.9*(y-5)^2/1)
if(log) return(log(out))
if(!log) return(out)})
}
