% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_samplik_2x.R
\name{lm.samplik.2x}
\alias{lm.samplik.2x}
\title{Calculates maximum sample likelihood estimators for linear regression}
\usage{
lm.samplik.2x(ys, xs, cutoffs, fh, fSRS = 0, in.s0, type = "SL",
  estimate.var = FALSE)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{xs}{matrix of covariate values (number of rows must equal n where n is length of ys).
Intercept is not assumed, so xs should contain a column of 1s if you want an intercept in the model.}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{fh}{Sampling fraction in each stratum for the stratified part of the sample}

\item{fSRS}{Sampling fraction of the SRSWOR component of the sample}

\item{in.s0}{A vector equal to 1 for units selected in the SRSWOR and to 0 for those
in the stratified SRSWOR. Default is that there are no SRSWOR units.}

\item{type}{If "SL", then the sample likelihood. If "CL" then the conditional
likelihood is returned, as defined by the first bracketed
expression in equation (2) of Zhou et al. (2002).}

\item{estimate.var}{If TRUE, variance is estimated using the observed information. Defaults to FALSE}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. If estimate.var=TRUE then variance estimates are provided.
}
\description{
This function fits a linear model to complex sample data by maximum sample likelihood
}
\details{
See Krieger and Pfefferman's paper.
The dependent variable is assumed to be normally distributed.
}
\examples{
data(population_example)
lm.samplik.2x(ys=sample.example$y,
              xs=cbind(1,sample.example$x1,sample.example$x2),
              cutoffs=c(0,2),fh=c(0.04,0.01,0.05))
}
