% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_WZ.R
\name{lm.WZ}
\alias{lm.WZ}
\title{Linear model fitting for stratified survey data by maximising the estimated likelihood of yu and xs}
\usage{
lm.WZ(ys, yr, xs, cutoffs, pi.s)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{yr}{vector of non-sample values of the dependent variable}

\item{xs}{matrix of covariate values (number of rows must equal n where n is length of ys).
Intercept is not assumed, so xs should contain a column of 1s if you want an intercept in the model.}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{pi.s}{n-vector (where n=length(y)) of probabilities of selection for the sample units}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. No variance estimates provided.
}
\description{
This function fits a linear model to survey data stratified by
the dependent variable, by maximising the estimated likelihood
from Weaver and Zhou (2012).
}
\examples{
data(population_example)
lm.WZ(ys=sample.example$y,yr=population.example$y[population.example$s.indicators==0],
xs=cbind(1,sample.example$x),cutoffs=c(1,2,3),pi.s=sample.example$pi)
}

