% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_Zhou_2x.R
\name{lm.Zhou.2x}
\alias{lm.Zhou.2x}
\title{Calculates maximum Zhou-likelihood estimators for linear regression of complex survey data under stratified SRSWOR}
\usage{
lm.Zhou.2x(ys, x1s, x2s, pi.s, cutoffs, in.s0)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{x1s}{sample values for covariate 1 (continuous)}

\item{x2s}{sample values for covariate 2 (binary)}

\item{pi.s}{n-vector (where n=length(y)) of probabilities of selection for the sample units}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata. NB ONLY H=3 IS ALLOWED!}

\item{in.s0}{A vector equal to 1 for units selected in the SRSWOR and to 0 for those
in the stratified SRSWOR. Default is that there are no SRSWOR units.}
}
\value{
a vector containing p estimated coefficients (where p=ncol(xs)) and the
estimated error standard deviation. No variance estimates provided.
}
\description{
This function fits a linear model to complex sample data by maximum Zhou-likelihood
}
\details{
Add some details later.
}
\examples{
data(population_example)
eg.ybar.pi <- sum(sample.example$y/sample.example$pi) / sum(1/sample.example$pi)
eg.s2.pi <- sum((sample.example$y-eg.ybar.pi)^2/sample.example$pi)/sum(1/sample.example$pi)
lm.Zhou.2x(ys=sample.example$y,x1s=sample.example$x1,x2s=sample.example$x2,
           pi.s=sample.example$pi,cutoffs=c(0,2))
}
