% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_points_in_intervals.R
\name{locate.points.in.intervals}
\alias{locate.points.in.intervals}
\title{Identifies which intervals some values lie within.}
\usage{
locate.points.in.intervals(y, cutoffs)
}
\arguments{
\item{y}{numeric vector}

\item{cutoffs}{A set of finite cutoffs defining the intervals.
If there are H cutoffs there are H+1 intervals}
}
\value{
a list with elements:
$interval.number: a vector of n integers (where n=length(y)) defining which
interval contains #' each element of y. A value of 1 corresponds to the
interval [-Inf, cutoff[1]), 2 corresponds to the interval
 [cutoff[1],cutoff[2]), ..., H+1 corresponds to [cutoff[H],Inf).
$lower.cutoffs: the lower bounds of the H+1 intervals
$upper.cutoffs: the upper bounds of the H+1 intervals
}
\description{
Identifies which intervals some values lie within.
}
\examples{
test <- locate.points.in.intervals(y=c(-100,-5:5),cutoffs=c(-10,-2,2,4,10))
cbind(y=c(-100,-5:5),interval=test$interval.number,
      lower=test$lower.cutoffs,
      upper=test$upper.cutoffs)
}
