% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods-PLM.R
\name{logspace}
\alias{logspace}
\title{Generate logarithmically spaced vector}
\usage{
logspace(d1, d2, n)
}
\arguments{
\item{d1}{first bound}

\item{d2}{second bound}

\item{n}{number of points}
}
\value{
a vector of n logarithmically spaced points between 10^d1 and 10^d2.
}
\description{
\code{logspace} generates n logarithmically spaced points between 10^d1 and
10^d2. The utility of this function is equivalent to logspace function in
matlab.
}
\examples{
logspace(-6,7,30)
}
