% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ls_2x.R
\name{Ls.2x}
\alias{Ls.2x}
\title{Calculates the sample likelihood}
\usage{
Ls.2x(ys, mean, sd, cutoffs, fh, fSRS = 0, in.s0, type = "SL",
  log = FALSE)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{mean}{vector of means of dependent variable}

\item{sd}{vector of standard deviations of dependent variable}

\item{cutoffs}{A vector of finite cutoffs defining the strata.
If there are H cutoffs there are H+1 strata.}

\item{fh}{Sampling fraction in each stratum for the stratified part of the sample}

\item{fSRS}{Sampling fraction of the SRSWOR component of the sample}

\item{in.s0}{A vector equal to 1 for units selected in the SRSWOR and to 0 for those
in the stratified SRSWOR. Default is that there are no SRSWOR units.}

\item{type}{If "SL", then the sample likelihood. If "CL" then the conditional
likelihood is returned, as defined by the first bracketed
expression in equation (2) of Zhou et al. (2002).}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}
}
\value{
a quantity proportional to the density of ys (or the log thereof)
}
\description{
This function calculates the sample likelihood (or rather something proportional to it)
}
\details{
See Krieger and Pfefferman's paper.
The dependent variable is assumed to be normally distributed
}
\examples{
data(population_example)
Ls.2x(mean=1+0.5*sample.example$x1-0.5*sample.example$x2,
      sd=1,ys=sample.example$y,cutoffs=c(0,2),fh=c(0.04,0.01,0.05),log=TRUE)
}
