% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ls.R
\name{Ls}
\alias{Ls}
\title{Calculates the sample likelihood}
\usage{
Ls(ys, mean, sd, pi.fn, log = FALSE, R = 100)
}
\arguments{
\item{ys}{vector of sample values of the dependent variable}

\item{mean}{vector of means of dependent variable}

\item{sd}{vector of standard deviations of dependent variable}

\item{pi.fn}{a function taking a single argument y and returning the probability
of selection}

\item{log}{If FALSE, the function returns the likelihood, otherwise the log-likelihood}

\item{R}{number of points used in quadrature calculation of integral}
}
\value{
a quantity proportional to the density of ys (or the log thereof)
}
\description{
This function calculates the sample likelihood (or rather something proportional to it)
}
\details{
See Krieger and Pfefferman's paper.
The dependent variable is assumed to be normally distributed
}
\examples{
data(population_example)
Ls(ys=sample.example$y,mean=5+0.5*sample.example$x,sd=1,log=TRUE,
pi.fn=function(y,log=FALSE){
out <- sqrt(0.1+0.9*(y-5)^2/1)
if(log) return(log(out))
if(!log) return(out)})
}

