% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_secondary.R
\docType{data}
\name{ods_data_secondary}
\alias{ods_data_secondary}
\title{Data example for the secondary analysis in ODS design}
\format{A matrix with 3000 rows and 7 columns: \describe{ \item{subj_ind}{An
  indicator variable for each subject: 1 = SRS, 2 = lowerODS, 3 = upperODS, 0
  = NVsample} \item{Y1}{primary outcome for which the ODS sampling scheme is
  based on} \item{Y2}{a secondary outcome} \item{X}{expensive exposure}
  \item{Z1}{a simulated covariate} \item{Z2}{a simulated covariate}
  \item{Z3}{a simulated covariate} }}
\source{
A simulated data set
}
\usage{
ods_data_secondary
}
\description{
Data example for the secondary analysis in ODS design
}
\keyword{datasets}
