% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_exp_adaptive.R
\name{p.s.exp.adaptive}
\alias{p.s.exp.adaptive}
\title{Calculates p(s) for a Poisson design with pi proportional to exp(c2*y/SY) for given c2}
\usage{
p.s.exp.adaptive(ys, yr, log = FALSE, specs = NULL)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. In this design
specs should be a 3-vector with named elements c2, n0 and return.what}
}
\value{
If specs$return.what=1, returns the probability that s was selected (or the log thereof).
        If specs$return.what=2, returns the probabilities of selection for each population unit
        If specs$return.what=3, returns a function pi.fn(y) which gives the unit probabilities of selection as a function of y
}
\description{
This function returns the probability p(s) of an entire sample being selected
for the Poisson sampling design where probabilities are proportional to exp(c2*y/SY).
SY is the population SD of Y, and constant of proportionality calculated such that
expected sample size is n0.
Probabilities of selection are truncated to be less than or equal to 1.
}
\details{
@details
 Add some details later.
}

