% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_exp.R
\name{p.s.exp}
\alias{p.s.exp}
\title{Calculates p(s) for a Poisson design with pi proportional to exp(c1+c2y) for given c1 and c2}
\usage{
p.s.exp(ys, yr, log = FALSE, specs = NULL)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. In this design
specs should be a 2-vector with named elements c1 and c2}
}
\description{
This function returns the probability p(s) of an entire sample being selected
for the Poisson sampling design where probabilities are proportional to exp(c1+c2y).
Probabilities of selection are truncated to be less than or equal to 1.
}
\details{
@details
 Note that expected sample size is not controlled.
}
