% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_extreme_quartiles.R
\name{p.s.extreme.quartiles}
\alias{p.s.extreme.quartiles}
\title{Calculates p(s) for the design selecting from quartiles 1 and 4}
\usage{
p.s.extreme.quartiles(ys, yr, log = F, specs = NULL)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design (not used for this design)}
}
\value{
The probability that s was selected (or the log thereof) multiplied by N/4-choose-n/4.
(The multiplication by N/4-choose-n/4 is to avoid very small values of p(s).)
}
\description{
This function returns the probability p(s) of an entire sample being selected
for the design with n/2 selected from first quartile and n/2 selected from fourth quartile
of the population values of the dependent variable
}
\details{
@details
 Insert some details here.
}

