% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_extremes.R
\name{p.s.extremes}
\alias{p.s.extremes}
\title{Calculates p(s) for the biggest and smallest design}
\usage{
p.s.extremes(ys, yr, log = FALSE, specs = NULL)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design (not used for this design)}
}
\value{
The probability that s was selected (or the log thereof). Either 1 or 0 for this design.
}
\description{
This function returns the probability p(s) of an entire sample being selected
 for the design where the largest n/2 and smallest n/2 values of the
 dependent variable are selected from the population
}
\details{
@details
 Insert some details here.
}
