% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_srswor.R
\name{p.s.srswor}
\alias{p.s.srswor}
\title{Calculates p(s) for simple random sampling without replacement (SRSWOR)}
\usage{
p.s.srswor(ys, yr, log = F, specs)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. Should contain n/N only.}
}
\value{
The probability that s was selected (or the log thereof) multiplied by a constant.
Because p(s) is equal either to a constant or to zero, the function just returns a value of p(s)=1 or 0.
}
\description{
This function returns the probability p(s) of an entire sample being selected
under SRSWOR
}
\details{
@details
 Add some details later.
}

