% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_strat.R
\name{p.s.strat}
\alias{p.s.strat}
\title{Calculates p(s) for the design stratified by y}
\usage{
p.s.strat(ys, yr, log = F, specs)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. Should be vector of
cutoffs (for H strata there should be a vector of H-1 cutoffs)}
}
\value{
The probability that s was selected (or the log thereof) multiplied by a constant.
}
\description{
This function returns the probability p(s) of an entire sample being selected
for stratified simple random sampling without replacement where strata are
defined by dividing y into intervals.
}
\details{
@details
 Add some details later.
}

