% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_stratq.R
\name{p.s.stratq}
\alias{p.s.stratq}
\title{Calculates p(s) for the design stratified by quartile}
\usage{
p.s.stratq(ys, yr, log = F, specs)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. Should be a 4-vector
containing the sample sizes in the quartiles 1-4.}
}
\value{
The probability that s was selected (or the log thereof) multiplied by a constant.
}
\description{
This function returns the probability p(s) of an entire sample being selected
for stratified simple random sampling without replacement where strata are
defined by the quartiles of the population values of the dependent variable
}
\details{
@details
 Add some details later.
}
