% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_s_ushaped.R
\name{p.s.ushaped}
\alias{p.s.ushaped}
\title{Calculates p(s) for a u-shaped design}
\usage{
p.s.ushaped(ys, yr, log = FALSE, specs = NULL)
}
\arguments{
\item{ys}{vector of the sample values of the dependent variable}

\item{yr}{vector of the non-sample values of the dependent variable}

\item{log}{If FALSE (the default), returns p(s). If TRUE, log(p(s)) is returned.}

\item{specs}{An object containing detailed specifications of the design. In this design
specs should be a 3-vector with named elements n0, tuner and return.what.}
}
\value{
If specs$return.what=1, returns the probability that s was selected (or the log thereof).
        If specs$return.what=2, returns the probabilities of selection for each population unit
        If specs$return.what=3, returns a function pi.fn(y) which gives the unit probabilities of selection as a function of y
}
\description{
This function returns the probability p(s) of an entire sample being selected
for the Poisson sampling design where probabilities are a u-shaped function of
y scaled such that the expected sample size is as specified.
}
\details{
@details
probability is proportional to sqrt(tuner+(1-tuner)*(y-meany.u)^2/vary.u)
truncated to lie be between 0.1*n0/N and 1-0.1*(N-n0)/N,
    or equivalently f0/10 and (f0+9)/10, where f0=n0/N,
}

