% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population.example}
\alias{population.example}
\title{Example population and sample data.}
\format{An object of class \code{data.frame} with 100 rows and 4 columns.}
\usage{
population.example
}
\description{
Example population and sample data.
}
\keyword{datasets}

