% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.example}
\alias{sample.example}
\title{Example population and sample data.}
\format{An object of class \code{data.frame} with 218 rows and 6 columns.}
\usage{
sample.example
}
\description{
Example population and sample data.
}
\keyword{datasets}
